requirements.txt
tensorflow~=2.1.0
matplotlib~=3.0
pytz
cptv~=1.5.0
opencv-python~=3.4
numpy~=1.18.5
scipy==1.4.1
python-dateutil
sklearn
tables~=3.4
h5py~=2.7
pyyaml>=4.2b1
pillow~=6.2.0
attrs~=19.1
filelock~=3.0.12
Astral~=1.10.1
timezonefinder~=4.1.0
pydbus~=0.6.0
PyGObject~=3.34.0
psutil~=5.6.7
portalocker~=1.5.2
toml~=0.10.0
cacophonyapi
